requirements.txt
Django
psycopg2-binary
docker-compose
pdbpp
python-vlc
black
mutagen